move TYPEsearchResult to their folders
