+++ VirtoCommerce.Storefront.Model/Cart/ShoppingCartSearchResult.cs
@@ -1,6 +1,6 @@
-using VirtoCommerce.Storefront.Model.Cart;
+using VirtoCommerce.Storefront.Model.Common;
 
-namespace VirtoCommerce.Storefront.Model.Common
+namespace VirtoCommerce.Storefront.Model.Cart
 {
     public class ShoppingCartSearchResult : GenericSearchResult<ShoppingCart>
     {
