--- VirtoCommerce.Storefront.Model/Common/QuoteSearchResult.cs
@@ -1,8 +1,0 @@
-using VirtoCommerce.Storefront.Model.Quote;
-
-namespace VirtoCommerce.Storefront.Model.Common
-{
-    public class QuoteSearchResult : GenericSearchResult<QuoteRequest>
-    {
-    }
-}
