--- VirtoCommerce.Storefront.Model/Common/SubscriptionSearchResult.cs
@@ -1,8 +1,0 @@
-using VirtoCommerce.Storefront.Model.Subscriptions;
-
-namespace VirtoCommerce.Storefront.Model.Common
-{
-    public class SubscriptionSearchResult : GenericSearchResult<Subscription>
-    {
-    }
-}
