--- VirtoCommerce.Storefront.Model/Common/UserSearchResult.cs
@@ -1,8 +1,0 @@
-using VirtoCommerce.Storefront.Model.Security;
-
-namespace VirtoCommerce.Storefront.Model.Common
-{
-    public class UserSearchResult : GenericSearchResult<User>
-    {
-    }
-}
