+++ VirtoCommerce.Storefront.Model/Inventory/FulfillmentCenterSearchResult.cs
@@ -1,6 +1,6 @@
-using VirtoCommerce.Storefront.Model.Inventory;
+using VirtoCommerce.Storefront.Model.Common;
 
-namespace VirtoCommerce.Storefront.Model.Common
+namespace VirtoCommerce.Storefront.Model.Inventory
 {
     public class FulfillmentCenterSearchResult : GenericSearchResult<FulfillmentCenter>
     {
