+++ VirtoCommerce.Storefront.Model/Order/CustomerOrderSearchResult.cs
@@ -1,6 +1,6 @@
-using VirtoCommerce.Storefront.Model.Order;
+using VirtoCommerce.Storefront.Model.Common;
 
-namespace VirtoCommerce.Storefront.Model.Common
+namespace VirtoCommerce.Storefront.Model.Order
 {
     public class CustomerOrderSearchResult : GenericSearchResult<CustomerOrder>
     {
