+++ VirtoCommerce.Storefront.Model/Quote/QuoteSearchResult.cs
@@ -1,0 +1,8 @@
+using VirtoCommerce.Storefront.Model.Common;
+
+namespace VirtoCommerce.Storefront.Model.Quote
+{
+    public class QuoteSearchResult : GenericSearchResult<QuoteRequest>
+    {
+    }
+}
