+++ VirtoCommerce.Storefront.Model/Security/UserSearchResult.cs
@@ -1,0 +1,8 @@
+using VirtoCommerce.Storefront.Model.Common;
+
+namespace VirtoCommerce.Storefront.Model.Security
+{
+    public class UserSearchResult : GenericSearchResult<User>
+    {
+    }
+}
