+++ VirtoCommerce.Storefront.Model/Subscriptions/SubscriptionSearchResult.cs
@@ -1,0 +1,8 @@
+using VirtoCommerce.Storefront.Model.Common;
+
+namespace VirtoCommerce.Storefront.Model.Subscriptions
+{
+    public class SubscriptionSearchResult : GenericSearchResult<Subscription>
+    {
+    }
+}
